Inheritance Changes
Player now has a method to get his feet z position.

BasicTileCube has a Box Collision. It has a function that gets the top
of the a cube tile that can be called by anything that extends
BasicTileCube.

NormalTile extends it, but is nothing but a tile that the player can
jump on.

PhaseableTile, UnstableTile, and WeakTile now all extend BasicTileCube.

UnstableTile and WeakTile now check to make sure the player is on top of
the platform before triggering their self destruct.