Damage Tiles Inheritance
Created a basicDamageTileCube that holds damage, duration of gadget
stun, and if the tile can gadget stun.

Created a method in the Player that when gadgets are disabled by event,
it creates a timer before the gadgets are enabled. If the event is
recalled before timer ends, the timer is refreshed.